+++ src/view/profile.js
@@ -1,18 +1,5 @@
-import {Rank, SpecialName} from "../utils.js";
+import {getRankOfUser} from "../utils.js";
 import AbstractView from "./abstract.js";
-
-const getRankOfUser = (cards) => {
-  const cardsLength = cards.filter((card) => card.isWatched).length;
-  let rank = ``;
-  if (cardsLength >= SpecialName.ONE && cardsLength <= SpecialName.TEN) {
-    rank = Rank.NOVICE;
-  } else if (cardsLength >= SpecialName.ELEVEN && cardsLength <= SpecialName.TWENTY) {
-    rank = Rank.FAN;
-  } else if (cardsLength >= SpecialName.TWENTY_ONE) {
-    rank = Rank.MOVIE_BUFF;
-  }
-  return rank;
-};
 
 const createProfileContentTemplate = (userRank) => {
   return `<section class="header__profile profile">
